+++ .travis_script.bash
@@ -3,7 +3,7 @@
 set -o errexit
 set -o nounset
 
-if [[ "$TRAVIS_BRANCH" == "release" && "$FAVORITE_CONFIG" != "yes" ]]
+if [[ "$TRAVIS_BRANCH" == "release" && "$RELEASE_CONFIG" != "yes" ]]
 then
   echo "This is a release branch and there's nothing this matrix element must do."
   exit 0
@@ -12,23 +12,6 @@ fi
 ponyc-test(){
   echo "Building and testing ponyc..."
   make CC="$CC1" CXX="$CXX1" test-ci
-}
-
-verify-changelog(){
-  echo "Building changelog tool..."
-  make CC="$CC1" CXX="$CXX1" && sudo make install
-
-  pushd /tmp
-
-  git clone "https://github.com/ponylang/pony-stable"
-  cd pony-stable && git checkout tags/0.0.1 && make && sudo make install && cd -
-
-  git clone "https://github.com/ponylang/changelog-tool"
-  cd changelog-tool && git checkout tags/0.2.0 && make && sudo make install && cd -
-
-  popd
-
-  changelog-tool verify CHANGELOG.md
 }
 
 ponyc-build-packages(){
@@ -58,50 +41,13 @@ ponyc-build-docs(){
 
 case "${TRAVIS_OS_NAME}:${LLVM_CONFIG}" in
 
-  "linux:llvm-config-3.7")
-    ponyc-test
-  ;;
-
-  "linux:llvm-config-3.8")
-    ponyc-test
-  ;;
-
   "linux:llvm-config-3.9")
-    # when FAVORITE_CONFIG stops matching part of this case, move this logic
-    if [[ "$TRAVIS_BRANCH" == "release" && "$TRAVIS_PULL_REQUEST" == "false" && "$FAVORITE_CONFIG" == "yes" ]]
+    # when RELEASE_CONFIG stops matching part of this case, move this logic
+    if [[ "$TRAVIS_BRANCH" == "release" && "$TRAVIS_PULL_REQUEST" == "false" && "$RELEASE_CONFIG" == "yes" ]]
     then
       ponyc-build-packages
       ponyc-build-docs
-    else
-      ponyc-test
-      verify-changelog
     fi
-  ;;
-
-  "linux:llvm-config-4.0")
-    ponyc-test
-  ;;
-
-  "linux:llvm-config-5.0")
-    ponyc-test
-  ;;
-
-  "osx:llvm-config-3.7")
-    echo "Running config=debug build..."
-    export config=debug
-    ponyc-test
-    echo "Running config=release build..."
-    export config=release
-    ponyc-test
-  ;;
-
-  "osx:llvm-config-3.8")
-    echo "Running config=debug build..."
-    export config=debug
-    ponyc-test
-    echo "Running config=release build..."
-    export config=release
-    ponyc-test
   ;;
 
   "osx:llvm-config-3.9")