+++ src/libponyrt/sched/scheduler.c
@@ -41,10 +41,6 @@ static bool use_yield;
 static mpmcq_t inject;
 static __pony_thread_local scheduler_t* this_scheduler;
 
-#if !defined(PLATFORM_IS_WINDOWS) && defined(USE_SCHEDULER_SCALING_PTHREADS)
-static pthread_cond_t sleep_cond;
-#endif
-
 /**
  * Gets the current active scheduler count
  */
@@ -126,9 +122,16 @@ static void wake_suspended_threads()
     {
       // in case the count changed between the while check and now
       if(get_active_scheduler_count() < scheduler_count)
+      {
         // send signal to wake up next scheduler thread available
-        ponyint_thread_wake(scheduler[get_active_scheduler_count()].tid,
-          scheduler[get_active_scheduler_count()].sleep_object);
+        if(ponyint_thread_wake(scheduler[get_active_scheduler_count()].tid,
+          scheduler[get_active_scheduler_count()].sleep_object))
+          // if there was an error waking the thread
+          // unlock the bool that controls modifying the active scheduler count
+          // variable.
+          atomic_store_explicit(&scheduler_count_changing, false,
+            memory_order_release);
+      }
       else
         // if there are no scheduler threads left to unlock
         // unlock the bool that controls modifying the active scheduler count
@@ -158,6 +161,11 @@ static bool read_msg(scheduler_t* sched)
         if(atomic_load_explicit(&detect_quiescence, memory_order_relaxed) &&
           (sched->block_count == get_active_scheduler_count()))
         {
+          // reset ack token count to 0 because dynamic scheduler scaling means
+          // that a new thread can suspend changing active_scheduler_count and
+          // we can think we've received enough acks when we really haven't
+          sched->ack_count = 0;
+
           // If we think all threads are blocked, send CNF(token) to everyone.
           send_msg_all_active(sched->index, SCHED_CNF, sched->ack_token);
         }
@@ -171,6 +179,13 @@ static bool read_msg(scheduler_t* sched)
         if(atomic_load_explicit(&detect_quiescence, memory_order_relaxed) &&
           (sched->block_count == get_active_scheduler_count()))
         {
+          // reset ack token count to 0 because dynamic scheduler scaling means
+          // that a new thread can wake up changing active_scheduler_count and
+          // then block causing block_count == active_scheduler_count for a
+          // second time and if we don't reset, we can think we've received
+          // enough acks when we really haven't
+          sched->ack_count = 0;
+
           // If we think all threads are blocked, send CNF(token) to everyone.
           send_msg_all_active(sched->index, SCHED_CNF, sched->ack_token);
         }
@@ -407,6 +422,11 @@ static pony_actor_t* steal(scheduler_t* sched)
     if (!block_sent)
     {
       uint32_t current_active_scheduler_count = get_active_scheduler_count();
+
+      // make sure thread scaling order is still valid. we should never be
+      // active if the active_scheduler_count isn't larger than our index.
+      pony_assert(current_active_scheduler_count > (uint32_t)sched->index);
+
       if (steal_attempts < current_active_scheduler_count)
       {
         steal_attempts++;
@@ -423,16 +443,14 @@ static pony_actor_t* steal(scheduler_t* sched)
           && !atomic_exchange_explicit(&scheduler_count_changing, true,
             memory_order_acquire))
         {
-          // let sched 0 know we're suspending
-          send_msg(sched->index, 0, SCHED_SUSPEND, 0);
-
-          // dtrace suspend notification
-          DTRACE1(THREAD_SUSPEND, (uintptr_t)sched);
-
           // decrement active_scheduler_count so other schedulers know we're
           // sleeping
           uint32_t sched_count = atomic_load_explicit(&active_scheduler_count,
             memory_order_relaxed);
+
+          // make sure the scheduler count didn't change
+          pony_assert(sched_count == current_active_scheduler_count);
+
           atomic_store_explicit(&active_scheduler_count, sched_count - 1,
             memory_order_relaxed);
 
@@ -441,20 +459,45 @@ static pony_actor_t* steal(scheduler_t* sched)
           atomic_store_explicit(&scheduler_count_changing, false,
             memory_order_release);
 
+          // let sched 0 know we're suspending only after decrementing
+          // active_scheduler_count to avoid a race condition between
+          // when we update active_scheduler_count and scheduler 0 processes
+          // the SCHED_SUSPEND message we send it. If we don't do this,
+          // and scheduler 0 processes the SCHED_SUSPEND message before we
+          // decrement active_scheduler_count, it could think that
+          // active_scheduler_count > block_count and not start the CNF/ACK
+          // process for termination and potentiall hang the runtime instead
+          // of allowing it to reach quiescence.
+          send_msg(sched->index, 0, SCHED_SUSPEND, 0);
+
+          // dtrace suspend notification
+          DTRACE1(THREAD_SUSPEND, (uintptr_t)sched);
+
           // sleep waiting for signal to wake up again
           ponyint_thread_suspend(sched->sleep_object);
+
+          bool scc = atomic_load_explicit(&scheduler_count_changing,
+            memory_order_acquire);
+
+          // make sure scheduler_count_changing is true
+          pony_assert(scc);
 
           // increment active_scheduler_count so other schedulers know we're
           // awake again
           sched_count = atomic_load_explicit(&active_scheduler_count,
             memory_order_relaxed);
+
+          // make sure the scheduler count is correct still
+          pony_assert((sched_count + 1) == current_active_scheduler_count);
+
           atomic_store_explicit(&active_scheduler_count, sched_count + 1,
             memory_order_relaxed);
 
           // unlock the bool that controls modifying the active scheduler count
           // variable. this is because the signalling thread locks the control
           // variable before signalling
-          atomic_store_explicit(&scheduler_count_changing, false,
+          scc = false;
+          atomic_store_explicit(&scheduler_count_changing, scc,
             memory_order_release);
 
           // dtrace resume notification
@@ -603,18 +646,13 @@ static void ponyint_sched_shutdown()
     // close wait event objects
     CloseHandle(scheduler[i].sleep_object);
 #elif defined(USE_SCHEDULER_SCALING_PTHREADS)
+    // destroy pthread condition object
+    pthread_cond_destroy(scheduler[i].sleep_object);
+    POOL_FREE(pthread_cond_t, scheduler[i].sleep_object);
     // set sleep condition object to NULL
     scheduler[i].sleep_object = NULL;
 #endif
   }
-
-#if !defined(PLATFORM_IS_WINDOWS) && defined(USE_SCHEDULER_SCALING_PTHREADS)
-  int ret;
-  // destroy pthread condition object
-  ret = pthread_cond_destroy(&sleep_cond);
-  // TODO: What to do if `ret` is a non-recoverable error?
-  (void) ret;
-#endif
 
   ponyint_pool_free_size(scheduler_count * sizeof(scheduler_t), scheduler);
   scheduler = NULL;
@@ -635,9 +673,9 @@ pony_ctx_t* ponyint_sched_init(uint32_t threads, bool noyield, bool nopin,
   if(threads == 0)
     threads = ponyint_cpu_count();
 
-  // If no minimum thread count is specified, use 1
+  // If no minimum thread count is specified, use # of threads
   if(min_threads == 0)
-    min_threads = 1;
+    min_threads = threads;
 
   // If minimum thread count is > thread count, cap it at thread count
   if(min_threads > threads)
@@ -654,22 +692,21 @@ pony_ctx_t* ponyint_sched_init(uint32_t threads, bool noyield, bool nopin,
   asio_cpu = ponyint_cpu_assign(scheduler_count, scheduler, nopin,
     pinasio);
 
-#if !defined(PLATFORM_IS_WINDOWS) && defined(USE_SCHEDULER_SCALING_PTHREADS)
-  // initialize pthread condition object
-  int ret = pthread_cond_init(&sleep_cond, NULL);
-#endif
-
   for(uint32_t i = 0; i < scheduler_count; i++)
   {
 #if defined(PLATFORM_IS_WINDOWS)
     // create wait event objects
     scheduler[i].sleep_object = CreateEvent(NULL, FALSE, FALSE, NULL);
 #elif defined(USE_SCHEDULER_SCALING_PTHREADS)
-    // if it failed, set `sleep_object` to `NULL` for error
+    // create pthread condition object
+    scheduler[i].sleep_object = POOL_ALLOC(pthread_cond_t);
+    int ret = pthread_cond_init(scheduler[i].sleep_object, NULL);
     if(ret != 0)
+    {
+      // if it failed, set `sleep_object` to `NULL` for error
+      POOL_FREE(pthread_cond_t, scheduler[i].sleep_object);
       scheduler[i].sleep_object = NULL;
-    else
-      scheduler[i].sleep_object = &sleep_cond;
+    }
 #else
     scheduler[i].sleep_object = PONY_SCHED_SLEEP_WAKE_SIGNAL;
 #endif
@@ -799,9 +836,24 @@ void ponyint_sched_maybe_wakeup()
     !atomic_exchange_explicit(&scheduler_count_changing, true,
     memory_order_acquire))
   {
-    // send signal to wake up next scheduler thread available
-    ponyint_thread_wake(scheduler[current_active_scheduler_count].tid,
-      scheduler[current_active_scheduler_count].sleep_object);
+    // in case the count changed between the while check and now
+    if(get_active_scheduler_count() < scheduler_count)
+    {
+      // send signal to wake up next scheduler thread available
+      if(ponyint_thread_wake(scheduler[get_active_scheduler_count()].tid,
+        scheduler[get_active_scheduler_count()].sleep_object))
+        // if there was an error waking the thread
+        // unlock the bool that controls modifying the active scheduler count
+        // variable.
+        atomic_store_explicit(&scheduler_count_changing, false,
+          memory_order_release);
+    }
+    else
+      // if there are no scheduler threads left to unlock
+      // unlock the bool that controls modifying the active scheduler count
+      // variable.
+      atomic_store_explicit(&scheduler_count_changing, false,
+        memory_order_release);
   }
 }